+++ tests/specs/spec.js
@@ -7,16 +7,21 @@ import fixtures from '../../fixtures/index.json'
 const { browsers = [], crawlers = [] } = fixtures
 
 const { equal, fail } = assert
+let spawn
 
 describe(
   'specs',
   () => {
+    beforeEach(() => {
+      spawn = isbot.spawn()
+    })
+
     it('should not break with empty input', () => {
-      equal(isbot(), false)
+      equal(spawn(), false)
     })
 
     it(`should return false for all ${browsers.length} browsers`, () => {
-      const recognised = browsers.filter(isbot)
+      const recognised = browsers.filter(spawn)
 
       recognised.length && fail([
         `Recognised as bots ${recognised.length} user agents:`,
@@ -25,92 +30,106 @@ describe(
     })
 
     it(`should return true for all ${crawlers.length} crawlers`, () => {
-      const unrecognised = crawlers.filter(ua => !isbot(ua))
+      const unrecognised = crawlers.filter(ua => !spawn(ua))
       unrecognised.length && fail([
         `Unrecognised as bots ${unrecognised.length} user agents:`,
         ...unrecognised.map(item => ` - ${item}`)
       ].join('\n'))
     })
 
-    describe('isbot.extend', () => {
+    describe('spawn.extend', () => {
       const useragent = 'Mozilla/5.0 (Linux) Randomagent/93.0'
       const rule = 'randomagent/\\d+\\.\\d+'
 
       it(`should not detect "${rule}" as bot`, () => {
-        assert(!isbot(useragent))
+        assert(!spawn(useragent))
       })
 
       it(`should detect "${rule}" as bot`, () => {
-        isbot.extend([rule])
-        assert(isbot(useragent))
+        spawn.extend([rule])
+        assert(spawn(useragent))
       })
 
       it('should not extend an existing item', () => {
-        isbot.extend([rule])
-        isbot.extend([rule])
-        isbot.extend([rule])
-        isbot.exclude([rule])
-        console.log(isbot.list)
-        assert(!isbot(useragent))
+        spawn.extend([rule])
+        spawn.extend([rule])
+        spawn.extend([rule])
+        spawn.exclude([rule])
+        console.log(spawn.list)
+        assert(!spawn(useragent))
       })
     })
 
-    describe('isbot.exclude', () => {
+    describe('spawn.exclude', () => {
       const useragent = 'Mozilla/5.0 (Macintosh; intel mac os x 10_14_6) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/88.0.4324.175 Safari/537.36 Chrome-Lighthouse'
       const rule = 'chrome-lighthouse'
 
       it(`should detect "${rule}" as bot`, () => {
-        assert(isbot(useragent))
+        assert(spawn(useragent))
       })
 
       it(`should not detect "${rule}" as bot`, () => {
-        isbot.exclude([rule])
-        assert(!isbot(useragent))
+        spawn.exclude([rule])
+        assert(!spawn(useragent))
       })
 
       it('should remain silent when excluding non existing filter', () => {
-        isbot.exclude(['something'])
+        spawn.exclude(['something'])
       })
     })
 
-    describe('isbot.find', () => {
+    describe('spawn.find', () => {
       it('should not break with empty input', () => {
-        equal(isbot.find(), null)
+        equal(spawn.find(), null)
       })
 
       it('should return null for non bot browser', () => {
-        equal(isbot.find('Mozilla/5.0 (Linux) Firefox/93.0'), null)
+        equal(spawn.find('Mozilla/5.0 (Linux) Firefox/93.0'), null)
       })
 
       it('should return the rule used to identify as bot', () => {
-        equal(isbot.find('Mozilla/5.0 (compatible; SemrushBot-SA/0.97; +http://www.semrush.com/bot.html)'), 'Bot')
+        equal(spawn.find('Mozilla/5.0 (compatible; SemrushBot-SA/0.97; +http://www.semrush.com/bot.html)'), 'Bot')
+      })
+
+      it('should be able to remove match using find', () => {
+        const ua = 'Mozilla/5.0 (Linux; Android 10; SNE-LX1 Build/HUAWEISNE-L21; wv) AppleWebKit/537.36 (KHTML, like Gecko) Version/4.0 Chrome/94.0.4606.71 Mobile Safari/537.36 hap/1079/huawei com.huawei.fastapp/11.4.1.310 com.frvr.worlds.quickapp/4.0.17 ({"packageName":"quickSearch","type":"other","extra":"{}"})'
+        equal(spawn(ua), true)
+        spawn.exclude(spawn.matches(ua))
+        equal(spawn(ua), false)
+      })
+
+      it('should clear all rules relevant to a user agent string', () => {
+        const ua = 'Mozilla/5.0 (Linux; Android 10; SNE-LX1 Build/HUAWEISNE-L21; wv) AppleWebKit/537.36 (KHTML, like Gecko) Spider/1.0 Robot/1.0 Search/1.0 Chrome/94.0.4606.71'
+        equal(spawn(ua), true)
+        spawn.clear(ua)
+        equal(spawn(ua), false)
       })
     })
 
-    describe('isbot.spawn', () => {
+    describe('spawn.spawn', () => {
       it('should spawn isbot with its own list', () => {
         const newUA = 'Mozilla/5.0 (Linux) NeoBrowser/93.0'
         const botUA = 'Mozilla/5.0 (compatible; SemrushBot-SA/0.97; +http://www.semrush.com/bot.html)'
-        const isbot2 = isbot.spawn(['neobrowser'])
-        assert(!isbot(newUA))
-        assert(isbot2(newUA))
-        assert(isbot(botUA))
-        assert(!isbot2(botUA))
+        const spawn2 = spawn.spawn(['neobrowser'])
+        assert(!spawn(newUA))
+        assert(spawn2(newUA))
+        assert(spawn(botUA))
+        assert(!spawn2(botUA))
       })
       it('should not affect each others lists', () => {
         const newUA = 'Mozilla/5.0 (Linux) NeoBrowser/93.0'
-        const isbot1 = isbot.spawn()
-        const isbot2 = isbot.spawn()
-        isbot1.extend(['neobrowser'])
-        assert(isbot1(newUA))
-        assert(!isbot2(newUA))
+        const spawn1 = spawn.spawn()
+        const spawn2 = spawn.spawn()
+        spawn1.extend(['neobrowser'])
+        assert(spawn1(newUA))
+        assert(!spawn2(newUA))
       })
       it('should spawn from instance\'s list', () => {
         const newUA = 'Mozilla/5.0 (Linux) NeoBrowser/93.0'
-        const isbot1 = isbot.spawn()
-        isbot1.extend(['neobrowser'])
-        const isbot2 = isbot1.spawn()
-        assert(isbot2(newUA))
+        const spawn1 = spawn.spawn()
+        spawn1.extend(['neobrowser'])
+        const spawn2 = spawn1.spawn()
+        assert(spawn2(newUA))
       })
     })
   }